Move evaluators to slice
